+++ src/util/is_table_sorted.py
